+++ miresearch/miresearch_main.py
@@ -74,6 +74,9 @@ groupA.add_argument('-RunPost', dest='subjRunPost',
 groupA.add_argument('-SubjInfo', dest='subjInfo', 
                     help='Print info for each subject', 
                     action='store_true')
+groupA.add_argument('-SubjInfoFull', dest='subjInfoFull', 
+                    help='Print full info for each subject', 
+                    action='store_true')
 
 # SEARCH ACTIONS
 groupA.add_argument('-SearchSeriesDesc', dest='searchSerDesc', 
@@ -85,6 +88,9 @@ groupA.add_argument('-SearchSeriesDesc', dest='searchSerDesc',
 groupA.add_argument('-SummaryCSV', dest='SummaryCSV', 
                     help='Write summary CSV file (give output file name)', 
                     type=str, nargs="*", default=None)
+groupA.add_argument('-Summary', dest='Summary', 
+                    help='Print summary of provided subjects to commandline (best with -sA option)', 
+                    action='store_true')
 
 # WATCH DIRECTORY
 groupA.add_argument('-WatchDirectory', dest='WatchDirectory', 
@@ -196,6 +202,13 @@ def runActions(args, extra_runActions=None):
                         print(f"{iSubj} has {list(res.values())}")
                 elif not args.QUIET:
                     print(f"{iSubj} does not exist at {iSubj.dataRoot}")
+        elif args.subjInfoFull:
+            for iSubj in subjList:
+                if iSubj.exists():
+                    if args.DEBUG:
+                        print(f"Info: {iSubj.subjID}...")
+                    iSubj.infoFull()
+                    
 
         # === SUBJECT GROUP ACTIONS ===
         # --- SummaryCSV ---
@@ -205,7 +218,12 @@ def runActions(args, extra_runActions=None):
                 if len(args.SummaryCSV) > 1:
                     print(f"  With tags: {args.SummaryCSV[1:]}")
             subjList.writeSummaryCSV(args.SummaryCSV[0], extra_series_tags=args.SummaryCSV[1:])
-
+        
+        # --- Summary ---
+        elif args.Summary:
+            if not args.QUIET:
+                print(f"Info: summary for {len(args.subjNList)} subjects at {args.dataRoot}")
+            print(subjList)
 
     ## WATCH DIRECTORY ##
     elif args.WatchDirectory is not None: