+++ CHANGELOG.md
@@ -8,3 +8,8 @@ All notable changes to this project will be documented in this file.
 - subject - add relative dicom filename to series info in meta.json. 
 - watchdog improved logging and error management
 
+## [0.1.6] - 2024-06-28
+### Added
+- Fix passing of SubjClass to correctly run post load actions if updating dicoms for a subject
+- Add extra info printing at cmdline 
+